Payement_app/Backend/requirements.txt
accelerate==1.10.0
aioconsole==0.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
asttokens==2.4.1
asyncpg==0.30.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backoff==2.2.1
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.17
botocore==1.39.17
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
coverage==7.8.0
cryptography==45.0.6
customs==0.5.0
daphne==4.2.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dirtyjson==1.0.8
distro==1.9.0
Django==5.2
django-appconf==1.1.0
django-cors-headers==4.7.0
django-coverage==1.2.4
djangorestframework==3.16.0
docling==2.46.0
docling-core==2.45.0
docling-ibm-models==3.9.0
docling-parse==4.2.2
durationpy==0.10
easyocr==1.7.2
ecdsa==0.19.1
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.34.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.14.0
groq==0.31.1
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==20.1.0
h11==0.16.0
hf-xet==1.1.8
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
incremental==24.7.2
invoke==2.2.0
ipykernel==6.29.5
ipython==8.27.0
ipython_pygments_lexers==1.0.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.75
langchain-google-genai==2.0.10
langchain-groq==0.3.8
langchain-postgres==0.0.15
langchain-text-splitters==0.3.9
langdetect==1.0.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.2
langgraph-swarm==0.0.14
langsmith==0.4.15
latex2mathml==3.78.0
lazy_loader==0.4
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.0
llama-index-cli==0.5.0
llama-index-core==0.13.6
llama-index-embeddings-langchain==0.4.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.0
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-chroma==0.5.3
llama-index-vector-stores-postgres==0.6.6
llama-index-workflows==1.3.0
llama-parse==0.6.54
lxml==5.3.0
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
mysqlclient==2.2.7
narwhals==2.3.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.1
openai==1.107.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
paramiko==2.12.0
parso==0.8.4
pexpect==4.9.0
pgvector==0.3.6
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
plum-dispatch==1.7.4
posthog==5.4.0
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.0
psycopg==3.2.9
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pylatexenc==2.10
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-jose==3.5.0
python-magic==0.4.27
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
rtree==1.4.1
s3transfer==0.13.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
semchunk==2.2.2
service-identity==24.2.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spire-presentation==10.7.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
sshtunnel==0.4.0
stack-data==0.6.3
streamlit==1.49.1
striprtf==0.0.26
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tifffile==2025.6.11
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
triton==3.4.0
Twisted==25.5.0
txaio==25.6.1
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
unstructured==0.18.13
unstructured-client==0.42.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.20.2
zope.interface==7.2
zstandard==0.24.0
paytmchecksum=1.7.0